% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox_negative.R
\name{boxcox_negative}
\alias{boxcox_negative}
\title{Apply a Box-Cox transformation for negative numbers}
\usage{
boxcox_negative(x, lambda1, lambda2)
}
\arguments{
\item{x}{Vector of numerics}

\item{lambda1}{Transformation exponent, varies from -5 to 5}

\item{lambda2}{Transformation exponent, varies from -5 to 5}
}
\value{
Vector \code{x} transformed by Box-Cox
}
\description{
Apply a Box-Cox transformation for negative numbers
}
