% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox_positive.R
\name{boxcox_positive}
\alias{boxcox_positive}
\title{Apply a Box-Cox transformation for positive numbers}
\usage{
boxcox_positive(x, lambda)
}
\arguments{
\item{x}{Vector of numerics}

\item{lambda}{Box-Cox transformation parameter, varies from -5 to 5}
}
\value{
Vector \code{x} transformed by Box-Cox
}
\description{
Apply a Box-Cox transformation for positive numbers
}
