% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_sample.R
\docType{data}
\name{gapminder_sample}
\alias{gapminder_sample}
\title{gapminder lifeExp data}
\format{tibble}
\source{
https://cran.r-project.org/web/packages/gapminder/gapminder.pdf
}
\usage{
gapminder_sample
}
\description{
An excerpt of the gapminder dataset. Only life expectancy values.
}
\keyword{datasets}
